+++ gulpfile.js
@@ -29,6 +29,7 @@ try {
   }))
 }
 
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const VERSION = require('./package.json').version
 const paths = {
   zip: [
@@ -42,6 +43,7 @@ const paths = {
     '!gulpfile.js',
     '!key.pem',
     '!android/**',
+    '!ios/**',
   ],
   views: './html/*.html',
   nativeHTML: './html/index.html',
@@ -114,9 +116,9 @@ const mochacss = function() {
     .pipe(gulp.dest('./dist/css/'))
 }
 
-const android = async function() {
+const native = async function() {
   const execa = (await import('execa')).execa
-  const {stdout} = await execa('node', ['node_modules/.bin/cap', 'sync'])
+  const {stdout} = await execa('cap', ['sync'])
   console.log(stdout)
 }
 
@@ -124,9 +126,11 @@ const mocha = gulp.parallel(mochajs, mochacss)
 
 const thirdparty = gulp.parallel(mocha)
 
-const main = gulp.series(html, locales, js, thirdparty, icons, android)
+const assets = gulp.parallel(html, thirdparty, locales, icons)
 
-const dev = gulp.series(html, thirdparty, locales, icons)
+const build = gulp.parallel(assets, js)
+
+const main = gulp.series(build, native)
 
 const zip = function() {
   return gulp
@@ -152,7 +156,7 @@ const crx = function() {
   )
 }
 
-const release = gulp.series(main, zip, xpi, crx)
+const release = gulp.series(main, gulp.parallel(zip, xpi), crx)
 
 const publish = gulp.series(main, zip, function() {
   return webstore
@@ -165,15 +169,15 @@ const publish = gulp.series(main, zip, function() {
 })
 
 const watch = function() {
-  let jsWatcher = gulp.watch(paths.js, dev)
+  let jsWatcher = gulp.watch(paths.js, assets)
   let viewsWatcher = gulp.watch(paths.views, html)
   let localeWatcher = gulp.watch(paths.locales, locales)
-  let androidWatcher = gulp.watch(paths.dist, android)
+  let nativeWatcher = gulp.watch(paths.dist, native)
 
   jsWatcher.on('change', onWatchEvent)
   viewsWatcher.on('change', onWatchEvent)
   localeWatcher.on('change', onWatchEvent)
-  androidWatcher.on('change', onWatchEvent)
+  nativeWatcher.on('change', onWatchEvent)
 
   webpack(devConfig).watch({}, (err, stats) => {
     if (err) {
@@ -186,9 +190,9 @@ const watch = function() {
   })
 }
 
-function onWatchEvent(event) {
+function onWatchEvent(path) {
   console.log(
-    'File ' + event.path + ' was ' + event.type + ', running tasks...'
+    'File ' + path + ' was changed, running tasks...'
   )
 }
 
@@ -197,11 +201,11 @@ exports.js = js
 exports.mocha = mocha
 exports.watch = watch
 exports.release = release
-exports.watch = gulp.series(dev, watch)
+exports.watch = gulp.series(main, watch)
 exports.publish = publish
-exports.dev = dev
+exports.build = build
 exports.locales = locales
-exports.android = android
+exports.native = native
 /*
  * Define default task that can be called by just running `gulp` from cli
  */