+++ src/ui/views/native/Home.vue
@@ -12,8 +12,9 @@ import { SplashScreen } from '@capacitor/splash-screen'
 import { SendIntent } from 'send-intent'
 import Controller from '../../../lib/Controller'
 import packageJson from '../../../../package.json'
-import { Storage } from '@capacitor/storage'
+import { Preferences as Storage } from '@capacitor/preferences'
 import { Http } from '@capacitor-community/http'
+import Logger from '../../../lib/Logger'
 
 export default {
   name: 'Home',
@@ -58,16 +59,21 @@ export default {
       if (result.text) {
         console.log(result.text)
         let url, title
-        if (result.text.includes('\n')) {
-          [title, , url] = result.text.split('\n', 3)
-        } else {
-          url = result.text
-          const response = await Http.get({url})
-          const parser = new DOMParser()
-          const document = parser.parseFromString(response.data, 'text/html')
-          const titleElement = document.getElementsByTagName('title')[0]
-          if (titleElement) {
-            title = titleElement.textContent
+        const match = result.text.match(/https?:\/\/\S*/i)
+        url = match[0]
+        title = result.text.substring(0, match.index) + result.text.substring(match.index + match[0].length)
+        title = title.replace('\n', ' ').trim()
+        if (!title) {
+          try {
+            const response = await Http.get({ url })
+            const parser = new DOMParser()
+            const document = parser.parseFromString(response.data, 'text/html')
+            const titleElement = document.getElementsByTagName('title')[0]
+            if (titleElement) {
+              title = titleElement.textContent
+            }
+          } catch (e) {
+            Logger.log('Failed to fetch shared URL')
           }
         }
         this.$router.push({