+++ src/ui/views/native/Tree.vue
@@ -1,6 +1,5 @@
 <template>
   <div
-    v-touch="{down: () => onSwipeDown()}"
     style="height:100%">
     <Drawer :visible.sync="drawer" />
     <v-app-bar
@@ -84,18 +83,35 @@
       </v-btn>
     </v-app-bar>
     <v-main>
+      <v-progress-linear
+        v-if="syncProgress"
+        :value="syncProgress * 100 || 0"
+        color="blue darken-1" />
+      <v-card>
+        <v-breadcrumbs
+          v-if="breadcrumbs.length > 1"
+          :items="breadcrumbs">
+          <template #item="{ item }">
+            <v-breadcrumbs-item @click="currentFolderId = item.id">
+              <template v-if="item.id === tree.id">
+                <v-icon>mdi-home</v-icon>
+              </template>
+              <template v-else>
+                {{ item.title }}
+              </template>
+            </v-breadcrumbs-item>
+          </template>
+        </v-breadcrumbs>
+      </v-card>
       <v-alert
         v-if="Boolean(syncError)"
         dense
         outlined
         text
         type="warning"
-        class="ma-1"
-        v-text="syncError" />
-      <v-progress-linear
-        v-if="syncProgress"
-        :value="syncProgress * 100 || 0"
-        color="blue darken-1" />
+        class="ma-1">
+        {{ syncError }}
+      </v-alert>
       <v-progress-circular
         v-if="loading"
         indeterminate
@@ -125,7 +141,7 @@
             </v-list-item-avatar>
 
             <v-list-item-content>
-              <v-list-item-title v-text="item.title" />
+              <v-list-item-title>{{ item.title }}</v-list-item-title>
               <v-list-item-subtitle v-if="item.type === 'bookmark'">
                 {{ item.url | hostname }}
               </v-list-item-subtitle>
@@ -216,6 +232,14 @@
           @click="addBookmark">
           <v-icon>mdi-star</v-icon>
         </v-btn>
+        <v-btn
+          color="blue darken-1"
+          dark
+          small
+          fab
+          @click="importBookmarks">
+          <v-icon>mdi-import</v-icon>
+        </v-btn>
       </v-speed-dial>
     </v-main>
 
@@ -247,6 +271,11 @@
       :display.sync="isEditingFolder"
       :tree="tree"
       @save="editFolder($event)" />
+    <DialogImportBookmarks
+      v-if="isImportingBookmarks"
+      :parent-folder="currentFolderId"
+      :display.sync="isImportingBookmarks"
+      :account-id="id" />
   </div>
 </template>
 
@@ -260,10 +289,11 @@ import { Bookmark, Folder } from '../../../lib/Tree'
 import { actions } from '../../store/definitions'
 import { App } from '@capacitor/app'
 import sortBy from 'lodash/sortBy'
+import DialogImportBookmarks from '../../components/native/DialogImportBookmarks'
 
 export default {
   name: 'Tree',
-  components: { FaviconImage, DialogEditBookmark, DialogEditFolder, Drawer },
+  components: { DialogImportBookmarks, FaviconImage, DialogEditBookmark, DialogEditFolder, Drawer },
   filters: {
     hostname(url) {
       return new URL(url).hostname
@@ -280,6 +310,7 @@ export default {
       currentlyEditedBookmark: null,
       isAddingBookmark: false,
       isAddingFolder: false,
+      isImportingBookmarks: false,
       fab: false,
       searchDebounceTimer: null,
       sortIcons: {
@@ -348,6 +379,13 @@ export default {
     },
     currentFolder() {
       return this.findItem(this.currentFolderId, this.tree)
+    },
+    breadcrumbs() {
+      const folders = [this.currentFolder]
+      while (folders[folders.length - 1 ].id !== this.tree.id) {
+        folders.push(this.findItem(folders[folders.length - 1 ].parentId, this.tree))
+      }
+      return folders.reverse()
     },
   },
   watch: {
@@ -458,6 +496,9 @@ export default {
     addBookmark() {
       this.isAddingBookmark = true
     },
+    importBookmarks() {
+      this.isImportingBookmarks = true
+    },
     createBookmark(props) {
       this.$store.dispatch(actions.CREATE_BOOKMARK, {
         accountId: this.id,
@@ -492,11 +533,6 @@ export default {
       this.currentAccount.data.syncing = 0.0001 // faaast
       this.$store.dispatch(actions.TRIGGER_SYNC, this.id)
     },
-    onSwipeDown() {
-      if (window.scrollY === 0) {
-        this.onTriggerSync()
-      }
-    }
   }
 }
 </script>